+++ Source/ACE/Entity/Landblock.cs
@@ -14,6 +14,8 @@ using ACE.Network.Enum;
 using ACE.Factories;
 using ACE.Entity.Enum;
 using ACE.DatLoader.FileTypes;
+using ACE.Network.GameMessages.Messages;
+using ACE.Entity.Enum.Properties;
 
 namespace ACE.Entity
 {
@@ -644,6 +646,53 @@ namespace ACE.Entity
                         player.Session.Network.EnqueueSend(identifyResponse);
                         break;
                     }
+                case GameActionType.Buy:
+                    {
+                        // todo: lots, need vendor list, money checks, etc.
+
+                        var money = new GameMessagePrivateUpdatePropertyInt(player.Session, PropertyInt.CoinValue, 4000);
+                        var sound = new GameMessageSound(player.Guid, Sound.PickUpItem, 1);
+                        var sendUseDoneEvent = new GameEventUseDone(player.Session);
+                        player.Session.Network.EnqueueSend(money, sound, sendUseDoneEvent);
+
+                        // send updated vendor inventory.
+                        player.Session.Network.EnqueueSend(new GameEventApproachVendor(player.Session, action.ObjectId));
+
+                        // this is just some testing code for now.
+                        foreach (ItemProfile item in action.ProfileItems)
+                        {
+                            // todo: something with vendor id and profile list... iid list from vendor dbs.
+                            // todo: something with amounts..
+
+                            if (item.Iid == 5)
+                            {
+                                while (item.Amount > 0)
+                                {
+                                    item.Amount--;
+                                    WorldObject loot = LootGenerationFactory.CreateTestWorldObject(5090);
+                                    LootGenerationFactory.AddToContainer(loot, player);
+                                    player.TrackObject(loot);
+                                }
+                                var rudecomment = "Who do you think you are, Johny Apple Seed ?";
+                                var buyrudemsg = new GameMessageSystemChat(rudecomment, ChatMessageType.Tell);
+                                player.Session.Network.EnqueueSend(buyrudemsg);
+                            }
+                            else if (item.Iid == 10)
+                            {
+                                while (item.Amount > 0)
+                                {
+                                    item.Amount--;
+                                    WorldObject loot = LootGenerationFactory.CreateTestWorldObject(30537);
+                                    LootGenerationFactory.AddToContainer(loot, player);
+                                    player.TrackObject(loot);
+                                }
+                                var rudecomment = "That smells awful, Enjoy eating it!";
+                                var buyrudemsg = new GameMessageSystemChat(rudecomment, ChatMessageType.Tell);
+                                player.Session.Network.EnqueueSend(buyrudemsg);
+                            }
+                        }
+                        break;
+                    }
                 case GameActionType.PutItemInContainer:
                     {
                         var playerGuid = new ObjectGuid(action.ObjectId);
@@ -777,6 +826,9 @@ namespace ACE.Entity
                                 (obj as Portal).OnCollide(player);
                             else if ((obj.DescriptionFlags & ObjectDescriptionFlag.Door) != 0)
                                 (obj as Door).OnUse(player);
+
+                            else if ((obj.DescriptionFlags & ObjectDescriptionFlag.Vendor) != 0)
+                                (obj as Vendor).OnUse(player);
 
                             // switch (obj.Type)
                             // {