+++ Source/ACE/Entity/WorldObject.cs
@@ -1,4 +1,4 @@
-using ACE.Entity.Enum;
+﻿using ACE.Entity.Enum;
 using ACE.Network;
 using ACE.Network.Enum;
 using ACE.Network.GameMessages.Messages;